Move with_buf_read into certificate_helpers, it is used there
